thin down command logic to a custom type
